style: 웹뷰 액티비티 제거
