+++ presentation/src/main/AndroidManifest.xml
@@ -10,8 +10,5 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity
-            android:name=".ui.CommonWebViewActivity"
-            android:exported="false" />
     </application>
 </manifest>
