--- presentation/src/main/java/com/mashup/alcoholfree/presentation/ui/CommonWebViewActivity.kt
@@ -1,63 +1,0 @@
-package com.mashup.alcoholfree.presentation.ui
-
-import android.annotation.SuppressLint
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import com.google.accompanist.web.WebContent
-import com.google.accompanist.web.WebView
-import com.google.accompanist.web.WebViewState
-import com.mashup.alcoholfree.presentation.ui.theme.AlcoholFreeAndroidTheme
-
-class CommonWebViewActivity : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContent {
-            AlcoholFreeAndroidTheme {
-                SulSulWebView(
-                    modifier = Modifier.fillMaxSize(),
-                    state = WebViewState(
-                        WebContent.Url(
-                            url = "https://www.naver.com",
-                            additionalHttpHeaders = emptyMap()
-                        )
-                    )
-                )
-            }
-        }
-    }
-}
-
-@Composable
-fun SulSulWebView(
-    modifier: Modifier = Modifier,
-    state: WebViewState,
-) {
-    WebView(
-        modifier = modifier,
-        state = state,
-    )
-}
-
-@SuppressLint("SetJavaScriptEnabled", "JavascriptInterface")
-@Composable
-fun SulSulWebView(
-    modifier: Modifier = Modifier,
-    state: WebViewState,
-    bridge: Any,
-    bridgeName: String,
-) {
-    WebView(
-        modifier = modifier,
-        state = state,
-        onCreated = { webView ->
-            with(webView) {
-                settings.javaScriptEnabled = true
-                addJavascriptInterface(bridge, bridgeName)
-            }
-        }
-    )
-}
